+++ src/theme/ThemeProvider.d.ts
@@ -1,0 +1,12 @@
+import {PureComponent} from 'react'
+import {Jss, SheetsRegistry, GenerateId} from 'jss'
+import {Theme} from './create-theme'
+
+export interface ThemeProviderProps {
+  theme?: Theme | ((theme: Theme) => Theme)
+  jss?: Jss
+  sheetsRegistry?: SheetsRegistry
+  generateId?: GenerateId
+}
+
+export class ThemeProvider extends PureComponent<ThemeProviderProps, {}> {}