+++ src/theme/create-theme.d.ts
@@ -1,6 +1,2 @@
 export type Theme = object
 export const createTheme: (props: {colors: object}) => Theme
-
-declare const theme: Theme
-
-export default theme